+++ database/factories/PostFactory.php
@@ -22,7 +22,13 @@ class PostFactory extends Factory
     public function definition()
     {
         return [
-            //
+            'title' => $this->faker->word,
+            'body' => []
         ];
+    }
+//  這樣可以給予 title 預設值
+    public function untitled()
+    {
+        return $this->state(['title' => 'untitled']);
     }
 }