+++ database/seeders/CommentSeeder.php
@@ -2,10 +2,15 @@
 
 namespace Database\Seeders;
 
+use App\Models\Comment;
+use Database\Seeders\Traits\DisableForeignKeys;
+use Database\Seeders\Traits\TruncateTable;
 use Illuminate\Database\Seeder;
 
 class CommentSeeder extends Seeder
 {
+    use TruncateTable, DisableForeignKeys;
+
     /**
      * Run the database seeds.
      *
@@ -13,6 +18,9 @@ class CommentSeeder extends Seeder
      */
     public function run()
     {
-        //
+        $this->disableForeignKey();
+        $this->truncate('comments');
+        Comment::factory(3)->create();
+        $this->enableForeignKey();
     }
 }