+++ CHANGELOG.md
@@ -6,10 +6,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.0.1] - 2021-08-25
+## [0.0.3] - 2021-08-26
 ### Added
-- 創建 Post、Comment 的 Model、Factory、Created、Seeder、Controller
+- 新增 Database Seeders
 
 ## [0.0.2] - 2021-08-25
 ### Added
 - 創建 Database 裡面的欄位。
+
+## [0.0.1] - 2021-08-25
+### Added
+- 創建 Post、Comment 的 Model、Factory、Created、Seeder、Controller
+
+
+
+